This is c code for binary search.
limitation in binary search is that you have to enter the sorted array to find the search element. sorted array in the sense it may be ascending or descending this code will give you output for ascending order you could change the if condition and work it for descending order.